error messages were ignoring their substitutions
Both of the following compile just fine:

  error $ printf "the number is %d" 42
  error "the number is %d" 42

But while the first throws "the number is 42", the second throws "the
number is %d"! The error was not caught by the type checker because
error has type "String -> a", so if a String is expected,

  error "the number is 42"

throws "the number is 42" and instantiates "a" to "String", and

  error "the number is %s" 42

throws "the number is %d" and instantiates "a" to "Int -> String", which
is not at all what you want.