Json objects should preserve the key order
Well, not according to the yaml documentation, they shouldn't, but it's
really annoying that the generated case classes have their fields in an
arbitrary order, so I'm modifying the aeson and yaml packages in order
to preserve the order in which the fields were listed in the file.

This commit only changes the type of Data.Aeson.Ordered.Object, breaking
everything. Subsequent commits will complete the refactoring by fixing
all the type errors.