new abstraction: Annotator
I added AnnotatedTree because I wanted to separate the task of
generating the different overlays into many much smaller tasks:
- annotating nodes with easy pieces of information (such as field names)
- combining simple annotations into more complex annotations
- once nodes are annotated with code, use it in the Generator

Different kinds of nodes require different kinds of annotations, so
AnnotatedTree has four type parameters corresponding to the four kinds
of node. Unfortunately this makes combining annotations a pain, because
its zipWith now takes four functions instead of one. In particular, we
can't use Applicative syntax :(

Even if we could, the Applicative instance wouldn't be lawful because we
can only meaningfully combine trees which have exactly the same shape.
And what would we even do with pure?

So instead of combining trees, let's combine Annotators! An annotator
computes an annotation for a node using only the information along the
path from the root to the node, and the information in the node's
children. Annotators can be combined using Applicative, and annotators
for child nodes can propagate their information up to their parents.
However, parent annotators cannot aggregate information from their child
nodes and then propagate that information down to their children,
because the result would depend on information which is not available to
annotators.